Went back to Integer in paging. Better logging when error in parsing json. Now handling enum in parsing json.

+++ core/src/main/java/com/jayway/forest/reflection/PagingSortingParameter.java
@@ -13,15 +13,15 @@ import java.util.List;
  * has not been done and it will handle sorting
  * and paging itself.
  */public interface PagingSortingParameter {
-    Long getPage();
+    Integer getPage();
 
-    Long getPageSize();
+    Integer getPageSize();
 
-    void setPageSize(Long pageSize);
+    void setPageSize(Integer pageSize);
 
-    void setTotalElements(Long totalElements);
+    void setTotalElements(Integer totalElements);
 
-    Long offset();
+    Integer offset();
 
     List<SortParameter> sortParameters();
